backend/requirements.render.txt
# Trading AI Platform - Backend Requirements for Render
# Standard OpenAI SDK (not emergentintegrations)

fastapi==0.115.0
uvicorn==0.34.0
python-dotenv==1.0.1
pymongo==4.9
bcrypt==4.2.0
PyJWT==2.11.0
python-jose==3.3.0
pydantic==2.9.2
python-multipart==0.0.9
openai==1.99.9
stripe==14.3.0
py-vapid==1.9.4
pywebpush==2.3.0
passlib==1.7.4
email-validator==2.3.0
